+++ src/test/java/com/milaboratory/mist/pattern/AndPatternTest.java
@@ -232,9 +232,9 @@ public class AndPatternTest {
                     randomGenerator.nextBoolean()).take() != null);
 
             for (Boolean fairSorting: new Boolean[] {false, true}) {
-                int errorScorePenalty = -randomGenerator.nextInt(50) - 1;
+                long errorScorePenalty = -randomGenerator.nextInt(50) - 1;
                 int maxOverlap = randomGenerator.nextInt(5) - 1;
-                int andPenaltyThreshold;
+                long andPenaltyThreshold;
                 boolean tooBigOverlap = false;
                 if (isMatchingPattern1) {
                     Match match1 = pattern1.match(targetQ).getBestMatch(fairSorting);
